+++ src/main/java/com/rozsa/crow/sound/SfxData.java
@@ -3,19 +3,16 @@ package com.rozsa.crow.sound;
 import com.rozsa.crow.game.api.Identifiable;
 
 public class SfxData implements Identifiable {
-    private int id;
-
+    private long id;
     private String name;
-
     private String soundFile;
-
     private int length;
 
-    public int getId() {
+    public long getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(long id) {
         this.id = id;
     }
